requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
sqlalchemy
pydantic

